requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
reportlab>=4.0.0
python-pptx>=0.6.23
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0
PyMuPDF>=1.23.0
pdf2image>=1.17.0
pytesseract>=0.3.13
opencv-python>=4.12.0
PyMuPDF4llm>=0.0.27
easyocr>=1.7.2
